backend/requirements.txt
# Core API framework
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0

# Task queue and caching
celery[redis]>=5.3.0,<6.0.0
redis>=4.5.0,<5.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
alembic>=1.12.0,<2.0.0

# Text extraction
PyMuPDF>=1.23.0,<2.0.0
python-docx>=1.1.0,<2.0.0
ebooklib>=0.18,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<5.0.0

# TTS and audio processing
TTS>=0.20.0,<1.0.0
pydub>=0.25.0,<1.0.0
soundfile>=0.12.0,<1.0.0
numpy>=1.22.0,<1.25.0

# Utilities
python-multipart>=0.0.6,<1.0.0
aiofiles>=23.2.0,<24.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0